SFDM-1569 add validation

+++ src/main/java/ru/beeline/techradar/service/CategoryService.java
@@ -2,16 +2,18 @@ package ru.beeline.techradar.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.beeline.techradar.controller.RequestContext;
 import ru.beeline.techradar.domain.Category;
 import ru.beeline.techradar.domain.TechCategory;
 import ru.beeline.techradar.dto.PatchCategoryDTO;
+import ru.beeline.techradar.dto.PostCategoryDTO;
 import ru.beeline.techradar.dto.PutTechCategoryDTO;
+import ru.beeline.techradar.exception.ForbiddenException;
 import ru.beeline.techradar.repository.CategoryRepository;
 import ru.beeline.techradar.repository.TechCategoryRepository;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @Transactional
@@ -30,23 +32,20 @@ public class CategoryService {
         return categoryRepository.findAll();
     }
 
-    public Category addCategory(Category category) {
-        Optional<Category> categoryOptional = categoryRepository.findByName(category.getName());
-        if (categoryOptional.isPresent()) {
-            return categoryOptional.get();
-        } else {
-            return categoryRepository.save(category);
+    public Category addCategory(PostCategoryDTO category) {
+        if (!RequestContext.getRoles().contains("ADMINISTRATOR")) {
+            throw new ForbiddenException("403 Forbidden.");
         }
+        return categoryRepository.findByName(category.getName())
+                .orElseGet(() -> categoryRepository.save(Category.builder().name(category.getName()).build()));
     }
 
     public void putCategory(PutTechCategoryDTO category) {
-        Category savedEntity = null;
-        Optional<Category> entityOptional = categoryRepository.findByName(category.getJoinCategoryName());
-        if (entityOptional.isPresent()) {
-            savedEntity = entityOptional.get();
-        } else {
-            savedEntity = addCategory(Category.builder().name(category.getJoinCategoryName()).build());
+        if (!RequestContext.getRoles().contains("ADMINISTRATOR")) {
+            throw new ForbiddenException("403 Forbidden.");
         }
+        Category savedEntity = categoryRepository.findByName(category.getJoinCategoryName())
+                .orElseGet(() -> addCategory(PostCategoryDTO.builder().name(category.getJoinCategoryName()).build()));
         List<TechCategory> techCategories = techCategoryRepository.findByCategory_IdIn(category.getJoinedCategoriesId());
         Category finalSavedEntity = savedEntity;
         techCategories.forEach(techCategory -> techCategory.setCategory(finalSavedEntity));
@@ -55,14 +54,22 @@ public class CategoryService {
     }
 
     public void patchCategory(String id, PatchCategoryDTO category) {
-        Category entity = categoryRepository.findById(Integer.parseInt(id)).get();
-        entity.setName(category.getNewName());
+        if (!RequestContext.getRoles().contains("ADMINISTRATOR")) {
+            throw new ForbiddenException("403 Forbidden.");
+        }
+        Category entity = categoryRepository.findById(Integer.parseInt(id))
+                .orElseThrow(() -> new IllegalArgumentException("Category with id=" + id + " not found."));
+        entity.setName(category.getName());
         categoryRepository.save(entity);
     }
 
     public void deleteCategory(String id) {
+        if (!RequestContext.getRoles().contains("ADMINISTRATOR")) {
+            throw new ForbiddenException("403 Forbidden.");
+        }
         List<TechCategory> techCategories = techCategoryRepository.findByCategory_IdIn(new ArrayList<>(Integer.parseInt(id)));
-        if(techCategories==null || techCategories.isEmpty());        {
+        if (techCategories == null || techCategories.isEmpty()) ;
+        {
             categoryRepository.deleteById(Integer.parseInt(id));
         }
         new RuntimeException("table tech_category have category_id = " + id);
